Pull MultiUpstreamRowReceiver out of RowMergers
Also adds some class-level javadocs.
